Move custom rules to rules folder to keep all rules in that one folder
