+++ src/cfnlint/custom_rules.py
@@ -9,19 +9,19 @@ SPDX-License-Identifier: MIT-0
 import logging
 import cfnlint
 import cfnlint.runner
-import cfnlint.customRules.Operators
-import cfnlint.customRules.Rule
+import cfnlint.rules.custom.Operators
+import cfnlint.rules.custom.Rule
 
 LOGGER = logging.getLogger(__name__)
 CUSTOM_RULES_FILE = None
-Operator = {'EQUALS': lambda x, y, z: cfnlint.customRules.Operators.equalsOp(x, y, z),
-            'NOT_EQUALS': lambda x, y, z: cfnlint.customRules.Operators.notEqualsOp(x, y, z),
-            '==': lambda x, y, z: cfnlint.customRules.Operators.equalsOp(x, y, z),
-            '!=': lambda x, y, z: cfnlint.customRules.Operators.notEqualsOp(x, y, z),
-            'IN': lambda x, y, z: cfnlint.customRules.Operators.InSetOp(x, y, z),
-            'NOT_IN': lambda x, y, z: cfnlint.customRules.Operators.NotInSetOp(x, y, z),
-            '>=': lambda x, y, z: cfnlint.customRules.Operators.greaterOp(x, y, z),
-            '<=': lambda x, y, z: cfnlint.customRules.Operators.lessOp(x, y, z)}
+Operator = {'EQUALS': lambda x, y, z: cfnlint.rules.custom.Operators.equalsOp(x, y, z),
+            'NOT_EQUALS': lambda x, y, z: cfnlint.rules.custom.Operators.notEqualsOp(x, y, z),
+            '==': lambda x, y, z: cfnlint.rules.custom.Operators.equalsOp(x, y, z),
+            '!=': lambda x, y, z: cfnlint.rules.custom.Operators.notEqualsOp(x, y, z),
+            'IN': lambda x, y, z: cfnlint.rules.custom.Operators.InSetOp(x, y, z),
+            'NOT_IN': lambda x, y, z: cfnlint.rules.custom.Operators.NotInSetOp(x, y, z),
+            '>=': lambda x, y, z: cfnlint.rules.custom.Operators.greaterOp(x, y, z),
+            '<=': lambda x, y, z: cfnlint.rules.custom.Operators.lessOp(x, y, z)}
 
 def check(template, rules, runner):
     """ Process custom rule file """
@@ -35,7 +35,7 @@ def check(template, rules, runner):
             line_number = 1
             for line in customRules:
                 LOGGER.debug('Processing Custom Rule Line %d', line_number)
-                rule = cfnlint.customRules.Rule.make_rule(line, line_number)
+                rule = cfnlint.rules.custom.Rule.make_rule(line, line_number)
                 if rule.valid and rule.resourceType[0] != '#':
                     try:
                         resource_properties = template.get_resource_properties([rule.resourceType])
@@ -45,7 +45,7 @@ def check(template, rules, runner):
                         matches.append(cfnlint.rules.Match(
                             1, 1,
                             1, 1,
-                            template.filename, cfnlint.customRules.Operators.CustomRule('E9999', 'Error'),
+                            template.filename, cfnlint.rules.custom.Operators.CustomRule('E9999', 'Error'),
                             str(rule.operator) + ' not in supported operators: ' + str(list(Operator.keys())) + ' at ' + str(line), None))
                 line_number += 1
         arg_matches = []
