+++ src/cfnlint/rules/custom/Operators.py
