+++ src/cfnlint/rules/custom/Rule.py
